git-gui: double click toggles stage and unstaged
I've been introducing a fair number of people to git-gui, and they usually try to double-click to stage files.
I have to tell them they have to click the icon. So, current behavior is not super intuitive. 

The change keeps the old behavior, and adds toggle via double click. Works great on macOS, the delay is short enough and it feels right.
I have not tested other platforms though.